# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Controller",
        "description": "Controller for basic functionality, such as validation and authorization.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/investment-account": {
            "get": {
                "tags": [
                    "Investment Accounts"
                ],
                "summary": "Get the authenticated user's investment account details",
                "operationId": "1ae4fcd7fe5f83f3e0db063abcd43b6a",
                "responses": {
                    "200": {
                        "description": "Investment account details retrieved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/investment-account/top-up": {
            "post": {
                "tags": [
                    "Investment Accounts"
                ],
                "summary": "Top up the investment account",
                "operationId": "53daedfd5bd87ffd9720a11c34233a50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Investment account topped up successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/investment-account/withdraw": {
            "post": {
                "tags": [
                    "Investment Accounts"
                ],
                "summary": "Withdraw funds from the investment account",
                "operationId": "fc47796c50424c67d526cda1d4e41b10",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funds withdrawn successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Insufficient funds"
                    },
                    "404": {
                        "description": "Investment account not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/investment-account/transactions": {
            "get": {
                "tags": [
                    "Investment Accounts"
                ],
                "summary": "List the last 10 transactions of the investment account",
                "operationId": "b8a36c19b7b8aca4fb6f5ad88b650c1a",
                "responses": {
                    "200": {
                        "description": "List of transactions retrieved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/campaigns/{campaignId}/invest": {
            "post": {
                "tags": [
                    "Investments"
                ],
                "summary": "Invest in a campaign",
                "operationId": "0e1ce92c79862db8034d49cabaf923cc",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "interest_rate"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 1000
                                    },
                                    "interest_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Investment successfully made"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Insufficient funds or other error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/investments": {
            "get": {
                "tags": [
                    "Investments"
                ],
                "summary": "List the investments of the authenticated user",
                "operationId": "b03d5b0b52a7eef15fa82f549f346632",
                "responses": {
                    "200": {
                        "description": "List of investments retrieved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/campaigns/{campaignId}/investments": {
            "get": {
                "tags": [
                    "Investments"
                ],
                "summary": "List investments for a specific campaign",
                "operationId": "7a1bc058b96d6f2ed9f53febab36de8c",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of investments retrieved"
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                }
            }
        },
        "/api/investments/{investmentId}/cancel": {
            "post": {
                "tags": [
                    "Investments"
                ],
                "summary": "Cancel an investment",
                "operationId": "fd011496ea1863be9075ea3b8df54df8",
                "parameters": [
                    {
                        "name": "investmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investment cancelled successfully"
                    },
                    "400": {
                        "description": "Invalid state or already cancelled"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel Passport OAuth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}